requirements.txt
toml==0.10.2
awscli==1.16.283
botocore==1.13.19
boto3==1.10.19
azure-storage-blob==12.10.0
